move logging out of get_games
